Making snake and apple objects more configurable (#3)
WHAT: Instead of passing different arguments to the constructor, we only expect one Options argument for both objects, with many of the arguments optional.

WHY: With this approach will be easier to configure other types of games. For instance, an online game happening on server might not need graphics component.